- login page uses config - refactored auth, config, and nav into separate modules
