+++ frontend/src/app/app.component.html
@@ -1,7 +1,7 @@
 <div class="app-container" [class.app-is-mobile]="mobileQuery.matches">
     <mat-toolbar color="primary">
         <button mat-icon-button (click)="snav.toggle()"><mat-icon>menu</mat-icon></button>
-        <a class="navbar-title" href="/"><h1 class="app-name">{{ config.navbarTitle }}</h1></a>
+        <a class="navbar-title" href="/"><h1 class="app-name">{{ config.appName }}</h1></a>
     </mat-toolbar>
 
     <mat-sidenav-container class="app-sidenav-container" [style.marginTop.px]="mobileQuery.matches ? 56 : 0">
