+++ frontend/src/app/app.module.ts
@@ -17,12 +17,12 @@ import { LoginComponent } from './login/login.component';
 import { PrivacyComponent } from './privacy/privacy.component';
 import { EventComponent } from './event/event.component';
 import {MediaMatcher} from '@angular/cdk/layout';
-import {ConfigService} from './config.service';
+import {ConfigService} from './config/config.service';
 import { SidenavComponent } from './sidenav/sidenav.component';
 import { ExportComponent } from './export/export.component';
-import {TokenInterceptor} from './token-interceptor.service';
-import { AuthService} from './auth.service';
-import {AuthGuardService} from './auth-guard.service';
+import {TokenInterceptor} from './auth/token-interceptor.service';
+import { AuthService} from './auth/auth.service';
+import {AuthGuardService} from './auth/auth-guard.service';
 
 @NgModule({
   declarations: [
