+++ frontend/src/app/config/gloves-config.ts
@@ -1,0 +1,8 @@
+import {Config} from './config';
+
+/**
+ * Configuration for Blue Gloves app.
+ */
+export const CONFIG: Config = new Config('Blue Gloves 4 All',
+    'www.gloveygloversonsgloveemporium.org',
+    '/assets/images/logos/gloves-logo.png');
