+++ frontend/src/app/config/h2ms-config.ts
@@ -1,0 +1,8 @@
+import {Config} from './config';
+
+/**
+ * Configuration for Hand Hygiene app.
+ */
+export const CONFIG: Config = new Config('H2MS',
+    'www.h2ms.org',
+    '/assets/images/logos/h2ms-logo.png');
