--- frontend/src/app/gloves-config.ts
@@ -1,6 +1,0 @@
-import {Config} from './config';
-
-/**
- * Configuration for Blue Gloves app.
- */
-export const CONFIG: Config = new Config('Blue Gloves 4 All', 'www.gloveygloversonsgloveemporium.org');
