--- frontend/src/app/h2ms-config.ts
@@ -1,6 +1,0 @@
-import {Config} from './config';
-
-/**
- * Configuration for Hand Hygiene app.
- */
-export const CONFIG: Config = new Config('H2MS', 'www.h2ms.org');
