+++ frontend/src/app/login/login.component.html
@@ -1,9 +1,9 @@
 <div class="login row">
     <div class="col-md-3"></div>
     <mat-card class="col-md-6">
-            <img mat-card-image src="assets/images/h2ms-logo-cropped.png">
+            <img mat-card-image src="{{config.logoURL}}">
             <mat-divider></mat-divider>
-            <mat-card-title>Welcome to H2MS!</mat-card-title>
+            <mat-card-title>Welcome to {{ config.appName }}!</mat-card-title>
         <mat-card-content class="text-center">
             <mat-form-field class="col">
                 <input matInput #email placeholder="Email">
