+++ frontend/src/app/login/login.component.ts
@@ -1,6 +1,8 @@
 import {Component, OnInit} from '@angular/core';
-import {AuthService} from '../auth.service';
+import {AuthService} from '../auth/auth.service';
 import {Router} from '@angular/router';
+import {ConfigService} from "../config/config.service";
+import {Config} from "../config/config";
 
 @Component({
   selector: 'app-login',
@@ -11,8 +13,10 @@ export class LoginComponent implements OnInit {
 
     hide = true;
     loginAttempts = 2;
+    config: Config;
 
-    constructor(private auth: AuthService, private router: Router) {
+    constructor(private auth: AuthService, private router: Router, private configService: ConfigService) {
+        this.config = configService.getConfig();
     }
 
     ngOnInit() {
