+++ frontend/src/app/sidenav/sidenav.component.ts
@@ -2,10 +2,10 @@ import {MediaMatcher} from '@angular/cdk/layout';
 import {ChangeDetectorRef, Component, OnDestroy, ViewChild} from '@angular/core';
 import {Location} from '@angular/common';
 import {NAV_ITEMS} from '../app-routing.module';
-import {NavItem} from '../nav-item';
+import {NavItem} from './nav-item';
 import {MatSidenav} from '@angular/material';
-import {ConfigService} from '../config.service';
-import {Config} from '../config';
+import {ConfigService} from '../config/config.service';
+import {Config} from '../config/config';
 
 @Component({
     selector: 'app-sidenav',
